+++ src/main/java/org/folio/linked/data/service/hub/HubService.java
@@ -1,0 +1,9 @@
+package org.folio.linked.data.service.hub;
+
+import org.folio.linked.data.domain.dto.ResourceResponseDto;
+
+public interface HubService {
+
+  ResourceResponseDto previewHub(String hubUri);
+
+}